+++ hawc/apps/assessment/api/__init__.py
@@ -1,4 +1,3 @@
 from .filters import *  # noqa: F401,F403
-from .pagination import *  # noqa: F401,F403
 from .permissions import *  # noqa: F401,F403
 from .viewsets import *  # noqa: F401,F403
