--- hawc/apps/assessment/api/pagination.py
@@ -1,5 +1,0 @@
-from rest_framework.pagination import PageNumberPagination
-
-
-class DisabledPagination(PageNumberPagination):
-    page_size = None
