+++ hawc/apps/common/api/__init__.py
@@ -1,6 +1,6 @@
 from .filters import *  # noqa: F401,F403
 from .mixins import *  # noqa: F401,F403
-from .pagination import PaginationWithCount  # noqa: F401,F403
+from .pagination import *  # noqa: F401,F403
 from .permissions import *  # noqa: F401,F403
 from .throttling import *  # noqa: F401,F403
 from .viewsets import *  # noqa: F401,F403
