+++ hawc/apps/common/api/pagination.py
@@ -1,6 +1,10 @@
 from rest_framework.pagination import PageNumberPagination
 
 
+class DisabledPagination(PageNumberPagination):
+    page_size = None
+
+
 class PaginationWithCount(PageNumberPagination):
     def get_paginated_response(self, data):
         response = super().get_paginated_response(data)
